move establishing of connection to App.tsx, assiging it to FirebaseContext for reuse across the app
