+++ src/Showings/Showings.tsx
@@ -1,31 +1,18 @@
-import { collection, getDocs, getFirestore, query } from "@firebase/firestore";
-import { initializeApp } from "firebase/app";
-import { useEffect, useState } from "react";
+import { getDocs, query } from "@firebase/firestore";
+import { useContext, useEffect, useState } from "react";
 import ShowingCard from "./ShowingCard";
-
-const firebaseConfig = {
-    apiKey: import.meta.env.VITE_FIREBASE_KEY,
-    authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-    projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-    storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-    messagingSenderId: import.meta.env.VITE_FIREBASE_SENDER_ID,
-    appId: import.meta.env.VITE_FIREBASE_APP_ID
-}
+import FirebaseContext from "../../hooks/FirebaseContext";
 
 function Showings() {
     const [showings, setShowings] = useState<any[]>([])
-    
-    async function getDocuments(){
-        const app = initializeApp(firebaseConfig)
-        const db = getFirestore(app)
+    const database = useContext(FirebaseContext)
 
-        const showingsRef = collection(db, "showings")
-        
-        const queryObject = query(showingsRef)
+    async function getDocuments(){        
+        const queryObject = query(database)
         const snapshot = await getDocs(queryObject)
 
         const fetchedShowings: any[] = []
-        snapshot.forEach((document)=>{
+        snapshot.forEach((document: any)=>{
             const showing = {
                 id: document.id,
                 ...document.data()
